+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   (which will be automatically converted to float32).
 - .isna() function now works on strings too.
 - `.save()` is now a method of `Frame` class.
+- Warnings now have custom display hook.
 
 
 #### Changed